requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
newsapi-python==0.2.6
pathspec==0.8.0
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.6.8
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
